+++ plugins/dates/scratch.js
@@ -2,17 +2,15 @@
 import nlp from '../../src/three.js'
 import spacetime from 'spacetime'
 
-// import datePlugin from './src/plugin.js'
-import datePlugin from './builds/compromise-dates.mjs'
+import datePlugin from './src/plugin.js'
+// import datePlugin from './builds/compromise-dates.mjs'
 nlp.plugin(datePlugin)
-// nlp.verbose('tagger')
+nlp.verbose('tagger')
 // nlp.verbose('date')
 
-const fmt = (iso) => (iso ? spacetime(iso).format('{day-short} {nice} {year}') : '-')
+const fmt = iso => (iso ? spacetime(iso).format('{day-short} {nice} {year}') : '-')
 
 // process.env.DEBUG_DATE = true
-
-
 
 // date issues:
 // 'the month before christmas' vs 'a month before christmas'
@@ -45,10 +43,14 @@ txt = `5th day of q1 2002`
 // txt = 'on april 22nd'
 txt = 'in basically one week from now'
 txt = 'go shopping with april'
+txt = 'between Sept and Oct 2008'
+txt = 'only in 2018 and 2020'
+txt = '2024/02/05 and 2024/03/09'
 
-nlp.verbose('tagger')
+// nlp.verbose('tagger')
 let doc = nlp(txt).debug()
-// console.log(doc.times().get())
+// doc.debug('dates')
+console.log(doc.dates().get())
 // doc.times().format('24h')
 // doc.debug()