+++ plugins/dates/src/plugin.js
@@ -4,20 +4,21 @@ import tags from './model/tags.js'
 import words from './model/words/index.js'
 import regex from './model/regex.js'
 import version from './_version.js'
-
-// import matches from './compute/matches.js'
+import debug from './debug.js'
 
 export default {
   tags,
   words,
   compute,
   api,
-  mutate: (world) => {
+  mutate: world => {
+    // add our regexes
     world.model.two.regexText = world.model.two.regexText || []
     world.model.two.regexText = world.model.two.regexText.concat(regex)
-    // net = net || methods.one.buildNet(matches, world)
-    // world.model.two.matches = world.model.two.matches.concat(matches)
+    // add our debug('dates') method
+    world.methods.one.debug = world.methods.one.debug || {}
+    world.methods.one.debug.dates = debug
   },
   hooks: ['dates'],
-  version
+  version,
 }