+++ src/1-one/output/methods/debug/index.js
@@ -1,0 +1,12 @@
+import clientSide from './client-side.js'
+import tags from './tags.js'
+import chunks from './chunks.js'
+import highlight from './highlight.js'
+
+const debug = {
+  tags,
+  clientSide,
+  chunks,
+  highlight,
+}
+export default debug