+++ tests/three/numbers/units.test.js
@@ -44,7 +44,6 @@ test('units-parse:', function (t) {
   t.end()
 })
 
-
 test('units-convert:', function (t) {
   let doc = nlp('we covered 3 km² in total')
   doc.numbers().toText()
@@ -53,7 +52,6 @@ test('units-convert:', function (t) {
   doc = nlp('we covered 3km² in total')
   doc.numbers().toText()
   t.equal(doc.text(), 'we covered three km² in total', here + 'no space unit')
-
 
   t.equal(nlp('44,000 ft').has(44000), true, here + '44,000 ft')
   t.equal(nlp('44,000ft').has('44,000'), true, here + '44,000')
@@ -64,6 +62,25 @@ test('units-convert:', function (t) {
   t.end()
 })
 
+test('isUnits:', function (t) {
+  let doc = nlp('i stored 300gb of text in 8mins')
+  let m = doc.numbers().isUnit('mins')
+  t.equal(m.text(), '8mins', 'str-min')
+
+  m = doc.numbers().isUnit(['mins', 'db', 'gb'])
+  t.deepEqual(m.out('array'), ['300gb', '8mins'], 'arr-input')
+
+  m = doc.numbers().isUnit({ gb: true, gbs: true })
+  t.deepEqual(m.out('array'), ['300gb'], 'obj-input')
+
+  m = doc.numbers().isUnit()
+  t.deepEqual(m.out('array'), [], 'empty-input')
+
+  doc = nlp('i stored 300gb on a 600 GB harddrive')
+  m = doc.numbers().isUnit('gb')
+  t.deepEqual(m.out('array'), ['300gb', '600'], 'unit-normalized')
+  t.end()
+})
 // test('implicit units', function (t) {
 //   let arr = [
 //     // ['99%', '99%'],